// src/generated/types.ts
export interface AcceptRejectPurchaseInvoiceResponse {
  invoice: ChangeStatusInvoiceResponse
  success: boolean
}

export interface AcceptRejectResponse {
  invoice: ChangeStatusInvoiceResponse
  success: boolean
  message?: string | null
  httpStatus?: string | null
}

export interface AddCirTicketRequest {
  cirInvoiceId?: string | null
  data?: string | null
  amount?: number | null
  userComment?: string | null
  resourceType: CirResourceType
  cirTicketCategory: CirTicketCategory
}

export interface AmountChangeDto {
  comments?: string | null
  amount: number
  cancelDate?: string | null
  creationDate?: string | null
  cancelComments?: string | null
  changedId: number
  id: number
}

export interface AssignmentDto {
  assignmentContractNr?: string | null
  assignmentDebtorName?: string | null
  assignmentDebtorCompanyNr?: string | null
  assignmentIdfNr?: string | null
  originalIdfNr?: string | null
}

export interface CancelInvoiceMessageDto {
  invoiceId: number
  cancelComments?: string | null
}

export interface CancellationDto {
  cancelDate?: string | null
  cancelledBy?: number | null
  reason?: string | null
}

export interface ChangeStatusInvoiceResponse {
  invoiceNumber?: string | null
  status: PurchaseInvoiceStatus
}

export interface ChangeUser {
  firstName?: string | null
  lastName?: string | null
}

export interface CirHistoryDto {
  comment?: string | null
  assignment: AssignmentDto
  amountChanges?: AmountChangeDto[] | null
  cancellation: CancellationDto
  settlements?: SettlementDto[] | null
}
// ...
export interface EFiscalizationFiscalBillNumbersForInvoiceRepresentation {
  fiscalBillNumber?: string | null
  transactionType: TransactionType
  sufTaxCode?: string | null
}
// ...
export interface InvoiceChangeDto {
  id: number
  propertyName?: string | null
  oldValue?: string | null
  newValue?: string | null
  dateChanged: string
  user: ChangeUser
  version: number
  serviceDesk: boolean
  ispName?: string | null
}
// ...
export interface InvoiceHistoryDto {
  invoiceId: number
  invoiceChanges?: InvoiceChangeDto[] | null
}

export interface InvoiceLinkDto {
  invoiceId?: string | null
  cirInvoiceId?: string | null
  invoiceNumber?: string | null
  sentToCir: boolean
  status: SalesInvoiceStatus
}

export interface InvoiceTotalVatPaymentsCalculationDomesticDto {
  vatPerRateCalculations?: InvoiceTotalVatPerRateCalculationDomesticDto[] | null
  totalVatPaymentsDomestic: number
}

export interface InvoiceTotalVatPerRateCalculationDomesticDto {
  vatRate: number
  baseSumForPaymentVatRate: number
  vatCategory?: string | null
  vatPaymentPerRate: number
}
// ...
export interface PurchaseFileDto {
  id: number
  name?: string | null
  extension?: string | null
  invoiceId: number
  mainPdf: boolean
  mainXml: boolean
  isFitekInZip: boolean
  isLink: boolean
  link?: string | null
  extendedPdf: boolean
}

export interface PurchaseInvoiceCustomFieldDto {
  customField: CustomFields
  value?: string | null
}

export interface PurchaseInvoiceDto {
  invoiceId: number
  senderApplicationId: number
  senderId: number
  receiverId: number
  publicPurchaseContractSignerId: number
  receiverUserAccountId?: number | null
  status: PurchaseInvoiceStatus
  salesStatus: SalesInvoiceStatus
  invoiceNumber?: string | null
  senderReceiverContractNumber?: string | null
  accountingDateUtc: string
  invoiceDateUtc?: string | null
  paymentDateUtc: string
  invoiceSentDateUtc: string
  referenceNumber?: string | null
  modelNumber?: string | null
  fineRatePerDay: number
  description?: string | null
  note?: string | null
  orderNumber?: string | null
  currencyName?: string | null
  discountPercentage: number
  discountAmount: number
  sumWithoutVat: number
  vatRate: number
  vatSum: number
  sumWithVat: number
  totalPayments?: number | null
  serviceId?: string | null
  invoiceFilePath?: string | null
  sender?: string | null
  receiver?: string | null
  publicPurchaseContractSigner?: string | null
  duplicate: number
  isDeleted: boolean
  files?: PurchaseFileDto[] | null
  customFields?: PurchaseInvoiceCustomFieldDto[] | null
  purchaseSelectedPrepaymentInvoices?: PurchaseSelectedPrepaymentInvoiceDto[] | null
  invoiceMessage?: string | null
  acceptRejectMessage?: string | null
  cirInvoiceId?: string | null
  cirAmountChangeId?: string | null
  cirStatus: CirInvoiceStatus
  cirHistory: CirHistoryDto
  cirAssignationHistory: InvoiceHistoryDto
  cirSettledAmount: number
  isCreditInvoice: boolean
  isDebitNote: boolean
  cancelInvoiceMessage?: string | null
  stornoNumber?: string | null
  isPrepaymentInvoice: boolean
  addVatRate?: boolean | null
  receiverCalculatedVatRate?: number | null
  totalToPay?: number | null
  vatPointDate: VatPointDates
  vatNumberFactoringCompany?: string | null
  factoringContractNumber?: string | null
  sourceInvoiceSelectionMode: SourceInvoiceSelectionMode
  indebtednessPeriodFromDate: string
  indebtednessPeriodToDate: string
  reductionDate?: string | null
  sourceInvoices?: InvoiceLinkDto[] | null
  sourceInvoicesOutOfEfaktura?: SourceInvoiceOutOfEFakturaDto[] | null
  invoiceTotalVatPaymentsCalculationDomestic: InvoiceTotalVatPaymentsCalculationDomesticDto
  isRelatedToInvoicesFromPreviosPeriod: boolean
  creditInvoices?: InvoiceLinkDto[] | null
  debitNotes?: InvoiceLinkDto[] | null
  ispAcceptedToRepresentCompany?: boolean | null
  ispName?: string | null
  lotNumber?: string | null
  internalRoutingNumber?: string | null
  receiverCalculatesVat: boolean
  objectIdentifier?: string | null
  createdFromXml?: boolean | null
  roundingAmount: number
  fiscalBillNumbersRepresentation?: EFiscalizationFiscalBillNumbersForInvoiceRepresentation[] | null
  automaticalAcceptance?: boolean | null
}

export interface PurchaseInvoiceOverviewDto {
  invoiceId: number
  globUniqId?: string | null
  documentNumber?: string | null
  documentType: DocumentType
  cirInvoiceId?: string | null
  status: PurchaseInvoiceStatus
  supplierName?: string | null
  supplierRegistrationNumber?: string | null
  supplierVatRegistrationNumber?: string | null
  amount: number
  sumWithoutVat: number
  vatAmount: number
  roundingAmount: number
  currency?: string | null
  deliveryDate?: string | null
  prepaymentDate?: string | null
  dueDate?: string | null
  dateOfReduction?: string | null
  sentDate: string
}

export interface PurchaseInvoicesDto {
  purchaseInvoiceIds?: number[] | null
}

export interface PurchaseInvoiceStatusChangeDto {
  eventId: number
  date?: string | null
  newInvoiceStatus: PurchaseInvoiceStatus
  purchaseInvoiceId: number
  comment?: string | null
  cirInvoiceId?: string | null
  subscriptionKey?: string | null
  stornoNumber?: string | null
  cirAssignmentChange: CirAssignmentChange
  isSigned?: boolean | null
  isAutoAssigned: boolean
}

export interface PurchaseSelectedPrepaymentInvoiceDto {
  invoiceId: number
  invoiceNumber?: string | null
  prepayedAmount: number
  prepayedVAT: number
  currency?: string | null
  paymentDate: string
  invoiceSentDateUtc: string
  contractId: number
  outOfEfaktura: boolean
}
// ...
export interface SettlementDto {
  settlementDate?: string | null
  amount: number
  comment?: string | null
}

export interface SimplePurchaseInvoiceDto {
  invoiceId: number
  globUniqId?: string | null
  comment?: string | null
  cirStatus: CirInvoiceStatus
  cirInvoiceId?: string | null
  version: number
  lastModifiedUtc: string
  cirSettledAmount: number
  vatNumberFactoringCompany?: string | null
  factoringContractNumber?: string | null
  cancelComment?: string | null
  stornoComment?: string | null
  status: PurchaseInvoiceStatus
}
// ...
export interface SourceInvoiceOutOfEFakturaDto {
  sourceInvoiceOutOfEFakturaId: number
  sourceInvoiceNumber?: string | null
  invoiceGlobUniqId?: string | null
}